+++ src/Language/Node/NodeInterface.php
@@ -3,6 +3,7 @@
 namespace Digia\GraphQL\Language\Node;
 
 use Digia\GraphQL\Language\Location;
+use Digia\GraphQL\Language\Visitor\VisitorInterface;
 
 interface NodeInterface
 {
@@ -26,4 +27,41 @@ interface NodeInterface
      * @return string
      */
     public function toJSON(): string;
+
+    /**
+     * @param VisitorInterface   $visitor
+     * @param null               $key
+     * @param NodeInterface|null $parent
+     * @param array              $path
+     * @param array              $ancestors
+     * @return NodeInterface|null
+     */
+    public function acceptVisitor(
+        VisitorInterface $visitor,
+        $key = null,
+        ?NodeInterface $parent = null,
+        array $path = [],
+        array $ancestors = []
+    ): ?NodeInterface;
+
+    /**
+     * @return NodeInterface|null
+     */
+    public function getAncestor(int $depth = 1): ?NodeInterface;
+
+    /**
+     * @param NodeInterface $node
+     * @return bool
+     */
+    public function determineIsEdited(NodeInterface $node): bool;
+
+    /**
+     * @return bool
+     */
+    public function isEdited(): bool;
+
+    /**
+     * @param array $properties
+     */
+    public function setVisitorProps(array $properties);
 }
