+++ src/Language/Node/VariableDefinitionsAwareInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Digia\GraphQL\Language\Node;
+
+interface VariableDefinitionsAwareInterface
+{
+    /**
+     * @return VariableDefinitionNode[]
+     */
+    public function getVariableDefinitions(): array;
+}
